+++ tests/test_meta_recommenders.py
@@ -8,59 +8,105 @@ from baybe.recommenders import (
     FPSRecommender,
     RandomRecommender,
     SequentialMetaRecommender,
-    StreamingSequentialMetaRecommender,
     TwoPhaseMetaRecommender,
 )
+from baybe.recommenders.meta.sequential import StreamingSequentialMetaRecommender
 from tests.conftest import select_recommender
 
 RECOMMENDERS = [RandomRecommender(), FPSRecommender(), BotorchRecommender()]
 
 
-def test_twophase_meta_recommender():
-    """The recommender switches the recommender at the requested point."""
+@pytest.mark.parametrize("remain_switched", [False, True])
+def test_twophase_meta_recommender(remain_switched):
+    """The recommender switches the recommender at the requested point and
+    remains / reverts the switch depending on the configuration."""  # noqa
+    # Cross the switching point forwards and backwards
+    switch_after = 3
+    training_sizes = [2, 3, 2]
+
+    # Recommender objects
     initial_recommender = RandomRecommender()
     subsequent_recommender = RandomRecommender()
-    switch_after = 3
     recommender = TwoPhaseMetaRecommender(
         initial_recommender=initial_recommender,
         recommender=subsequent_recommender,
         switch_after=switch_after,
+        remain_switched=remain_switched,
     )
-    for training_size in range(6):
-        rec = select_recommender(recommender, training_size)
+
+    # Query the meta recommender
+    switch_point_passed = False
+    for n_data in training_sizes:
+        rec = select_recommender(recommender, n_data)
         target = (
-            initial_recommender
-            if training_size < switch_after
-            else subsequent_recommender
+            subsequent_recommender
+            if (n_data >= switch_after) or (switch_point_passed and remain_switched)
+            else initial_recommender
         )
+        if not switch_point_passed and n_data >= switch_after:
+            switch_point_passed = True
         assert rec is target
 
 
-@pytest.mark.parametrize("mode", ["raise", "reuse_last", "cyclic"])
-@pytest.mark.parametrize("recommenders", [RECOMMENDERS])
-def test_sequential_meta_recommender(recommenders, mode):
+@pytest.mark.parametrize(
+    ("cls", "mode"),
+    [
+        (SequentialMetaRecommender, "raise"),
+        (SequentialMetaRecommender, "reuse_last"),
+        (SequentialMetaRecommender, "cyclic"),
+        (StreamingSequentialMetaRecommender, None),
+    ],
+)
+def test_sequential_meta_recommender(cls, mode):
     """The recommender provides its recommenders in the right order."""
-    meta_recommender = SequentialMetaRecommender(recommenders=recommenders, mode=mode)
+    # Create meta recommender
+    if cls == SequentialMetaRecommender:
+        meta_recommender = SequentialMetaRecommender(
+            recommenders=RECOMMENDERS, mode=mode
+        )
+    else:
+        meta_recommender = StreamingSequentialMetaRecommender(
+            recommenders=(r for r in RECOMMENDERS)  # <-- generator comprehension
+        )
+
     training_size = 0
 
     # First iteration over provided recommender sequence
-    for reference in recommenders:
+    for reference in RECOMMENDERS:
         training_size += 1
 
         # The returned recommender coincides with what was put in
         recommender = select_recommender(meta_recommender, training_size)
         assert recommender is reference
 
+        # If the current recommender was not used (by recommending via the meta
+        # recommender), a subsequent call returns the same recommender
+        recommender = select_recommender(meta_recommender, training_size + 1)
+        assert recommender is reference
+
+        # Pretend the previous data size increase did not happen
+        meta_recommender._n_last_measurements = training_size
+
+        # Pretend the recommender was used
+        meta_recommender._was_used = True
+
         # Selection with unchanged training size yields again the same recommender
         recommender = select_recommender(meta_recommender, training_size)
         assert recommender is reference
 
         # Selection with smaller training size raises an error
-        with pytest.raises(RuntimeError):
+        with pytest.raises(
+            RuntimeError,
+            match=f"decreased from {training_size} to {training_size-1}",
+        ):
             select_recommender(meta_recommender, training_size - 1)
 
+    # For streaming recommenders, no second iteration is possible
+    if cls == StreamingSequentialMetaRecommender:
+        return
+
     # Second iteration over provided recommender sequence
-    for cycled in recommenders:
+    for cycled in RECOMMENDERS:
         training_size += 1
 
         if mode == "raise":
@@ -71,37 +117,12 @@ def test_sequential_meta_recommender(recommenders, mode):
         elif mode == "reuse_last":
             # The last recommender is selected repeatedly
             recommender = select_recommender(meta_recommender, training_size)
-            assert recommender == recommenders[-1]
+            assert recommender == RECOMMENDERS[-1]
 
         elif mode == "cyclic":
             # The selection restarts from the first recommender
             recommender = select_recommender(meta_recommender, training_size)
             assert recommender == cycled
 
-
-@pytest.mark.parametrize(
-    "recommenders",
-    [
-        RECOMMENDERS,  # list
-        (rec for rec in RECOMMENDERS),  # generator
-    ],
-)
-def test_streaming_sequential_meta_recommender(recommenders):
-    """The recommender provides its recommenders in the right order."""
-    meta_recommender = StreamingSequentialMetaRecommender(recommenders=recommenders)
-    training_size = 0
-
-    for reference in RECOMMENDERS:
-        training_size += 1
-
-        # The returned recommender coincides with what was put in
-        recommender = select_recommender(meta_recommender, training_size)
-        assert recommender is reference
-
-        # Selection with unchanged training size yields again the same recommender
-        recommender = select_recommender(meta_recommender, training_size)
-        assert recommender is reference
-
-        # Selection with smaller training size raises an error
-        with pytest.raises(RuntimeError):
-            select_recommender(meta_recommender, training_size - 1)
+        # Pretend the recommender was used
+        meta_recommender._was_used = True